[Add Pet] refactor to improve readability on main 'add' page
- flagged possible sections causing error
